+++ client/scrapescholar_client/__tests__/helperFunctions/accordianContainsSources.tsx
@@ -1,0 +1,13 @@
+import { render, screen, fireEvent, waitFor } from '@testing-library/react';
+
+const accordianContainsSources = () =>{
+
+        const scienceDirectChecklist = screen.getByText('ScienceDirect');
+        const scopusChecklist = screen.getByText('Scopus');
+        expect(scienceDirectChecklist).toBeInTheDocument();
+        expect(scopusChecklist).toBeInTheDocument();
+      
+}
+
+
+export default accordianContainsSources;
